unit tests
We implement unit tests by first defining a test script which asserts
the expected replies from the Brewery Lambda function.
Then we set up a null resource with a local-exec provisioner to
run the unit tests during Terraform apply, and a test-runner to
periodically run the tests after the application has been deployed.
We add the necessary minitest gem in a just-in-time Gemfile. If the
Gemfile were to ever become more complex, we would split it out
into an external file.
These changes require Bundler to be installed on the development machine.